requirements.txt
httpx[http2]
pandas
numpy
matplotlib
seaborn
scipy
pyyaml
tqdm
Jinja2
scikit-learn